requirements.txt
# API/Scraping
praw==7.8.1
tweepy==4.15.0
beautifulsoup4==4.13.3
requests==2.32.0
scrapy==2.12.0

# Data handling
pandas==2.2.3
pymongo==4.11.3

# Utils
pyyaml==6.0.2
python-dotenv==1.1.0

# Testing
pytest==8.3.5
pytest-mock==3.14.0

# Code quality
black==25.1.0
flake8==7.2.0

# Transformation
nltk>=3.9.1

# Modeling
scikit-learn>=1.6.0